% Auto-generated: do not edit by hand
\name{''ChartJs}

\alias{''ChartJs}

\title{ChartJs component}

\description{
This component renders ChartJs React component inside Dash App.
}

\usage{
''ChartJs(id=NULL, data=NULL, options=NULL, type=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. The data object that is passed into the Chart.js chart}

\item{options}{Named list. The options object that is passed into the Chart.js chart}

\item{type}{Character. Chart.js chart type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

