% Auto-generated: do not edit by hand
\name{''ChartJs}

\alias{''ChartJs}

\title{ChartJs component}

\description{
This component renders ChartJs React component inside Dash App.
}

\usage{
''ChartJs(id=NULL, clickData=NULL, customJSFunctions=NULL,
customPlugins=NULL, data=NULL, options=NULL, redraw=NULL,
style=NULL, toolbox=NULL, type=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{clickData}{Named list. clickData returns the datasetIndex and index of data point clicked.}

\item{customJSFunctions}{Named list. Write custom JS functions in dict format.}

\item{customPlugins}{Named list. Write Plugins in dict format.}

\item{data}{Named list. The data object that is passed into the Chart.js chart.}

\item{options}{Named list. The options object that is passed into the Chart.js chart.}

\item{redraw}{Logical. Teardown and redraw chart on every update.}

\item{style}{Named list. Defines CSS styles which will override styles previously set.}

\item{toolbox}{Logical. Toolbox with reset and download buttons for chart.}

\item{type}{Character. Chart.js chart type.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

